ImportData: move opendatasoft logic somewhere else
